% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpedigree.R
\name{ggpedigree}
\alias{ggpedigree}
\title{ggpedigree: Pedigree plotting tool for complex pedigrees.}
\usage{
ggpedigree(
  .data,
  ids,
  mothers,
  fathers,
  cohort,
  sex,
  sex_code = NULL,
  id_labels = FALSE,
  remove_singletons = TRUE,
  plot_unknown_cohort = TRUE,
  randomise_x_coordinates = FALSE,
  print_cohort_labels = TRUE,
  return_plot_tables = FALSE,
  line_col_mother = "#E41A1C",
  line_col_father = "#377EB8",
  line_alpha = 0.3,
  point_size = 1,
  point_colour = "black",
  point_alpha = 1
)
}
\arguments{
\item{.data}{an optional data frame object with all the pedigree information}

\item{ids}{a column of .data or a vector of individual identifiers}

\item{mothers}{A column of .data or a vector of mothers corresponding to ids. Missing values are 0
or NA.}

\item{fathers}{A column of .data or a vector of fathers corresponding to ids. Missing values are 0
or NA.}

\item{cohort}{integer. Default NULL. A column of .data or an optional vector assigning a cohort to
each id. If NULL, then \code{kinship2::kindepth} is used to assign cohorts to
ids.}

\item{sex}{integer or character. Default NULL. An optional column of .data or a vector assigning a
sex to each id. This can be any value, but the first level numerically or
alphabetically (e.g. 0 or "F") will be plotted with a circle (traditionally
denoting a female) and the second level will be plotted with a square
(traditionally denoting a male). Any negative values, NA values, or third
alphabetically values will be plotted with a triangle. NOTE: These can be
overridden by specifying \code{sex_female} or \code{sex_male} values.}

\item{sex_code}{Default NULL. A vector of length 2, indicatinging the value
used in \code{sex} for females and males respectively. Females will be plotted
as circles and males as squares.}

\item{id_labels}{logical. Default FALSE. Print the ids on the pedigree plot.}

\item{remove_singletons}{logical. Default TRUE. Remove ids with no relatives
i.e., no offspring or parents assigned.}

\item{plot_unknown_cohort}{logical. Default TRUE. Plots ids of unknown
cohorts. These are plotted in an "Unknown" cohort at the top of the
pedigree. Be aware that any mothers and fathers of these individuals will
be plotted below them.}

\item{randomise_x_coordinates}{logical. Default TRUE. Randomise the position of
individuals within each cohort.}

\item{print_cohort_labels}{logical. Default TRUE. Prints cohort levels on the
left hand side of plot.}

\item{return_plot_tables}{logical. Default FALSE. Returns an object with the
line and point data used for the plot, but the plot will not be generated}

\item{line_col_mother}{Default = "#E41A1C". Line colour for maternal links.}

\item{line_col_father}{Default = "#377EB8". Line colour for paternal links.}

\item{line_alpha}{Default = 0.3. Line alpha (transparency) value for maternal
and paternal links.}

\item{point_size}{Default = 1. Point size for ids.}

\item{point_colour}{Default = "black". Point colour for ids.}

\item{point_alpha}{Default = 1. Point alpha for ids.}
}
\description{
ggpedigree: Pedigree plotting tool for complex pedigrees.
}
\examples{
data(gryphons)
pedigree <- fix_ped(gryphons[, 1:3])

## draw the gryphon pedigree by pedigree depth
ggpedigree(pedigree)
\dontrun{
# specifying the column names for id, mother and father
ggpedigree(pedigree, id, dam, sire)

# with cohort and sex
ggpedigree(gryphons, cohort = cohort, sex = sex, sex_code = c(0, 1))
}
}
\keyword{plot}
