% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gryphons}
\alias{gryphons}
\title{Example dataset for pedtricks examples and tutorial}
\format{
An object of class \code{data.frame} with 4918 rows and 9 columns.
}
\usage{
gryphons
}
\description{
This contains pedigree and life history data of a fictional
population.  The data are relevant to power and sensitivity
analyses for quantitative genetic studies of natural populations.
}
\details{
\describe{
\item{id}{what}
\item{dam}{what}
\item{sire}{what}
\item{sex}{what}
\item{cohort}{what}
\item{fatherErrorProb}{what}
\item{fatherSampledProb}{what}
\item{firstReproCohort}{what}
\item{lastReproCohort}{what}
}
}
\keyword{dataset}
