% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedantics-deprecated.R
\name{pedantics-defunct}
\alias{pedantics-defunct}
\title{Defunct functions from package \pkg{pedantics}.}
\description{
The functions listed below have not been transferred from pedantics to
pedtricks with no short-term plan to implement them since they were relying
on C code that need to be heavily updated.
}
\section{\code{rpederr}}{
 Permutes a pedigree to create a plausible complete pedigree
}

\section{\code{fpederr}}{
 Simulates a pedigree with errors and missing data from a complete pedigree.
}

\keyword{internal}
