% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_ped.R, R/draw_ped.R, R/fix_ped.R,
%   R/genome_sim.R, R/micro_sim.R, R/ped_stats.R, R/ped_stats_summary.R,
%   R/pedantics-deprecated.R, R/phen_sim.R
\name{makemakePedigreeFactor}
\alias{makemakePedigreeFactor}
\alias{makePedigreeNumeric}
\alias{drawPedigree}
\alias{fixPedigree}
\alias{genomesim}
\alias{microsim}
\alias{pedigreeStat}
\alias{pedStatSummary}
\alias{pedantics-deprecated}
\alias{phensim}
\title{Deprecated functions from package \pkg{pedantics}.}
\usage{
makemakePedigreeFactor(id, sire, dam, key = NULL)

makePedigreeNumeric(id, sire, dam, missingVal = NA)

drawPedigree(
  Ped,
  cohorts = NULL,
  sex = NULL,
  dat = NULL,
  dots = "n",
  plotfull = "y",
  writeCohortLabels = "n",
  links = "all",
  sexInd = c(0, 1),
  dotSize = 0.001,
  dataDots = "n",
  dataDots.cex = 2,
  cohortLabs.cex = 1,
  retain = "informative",
  focal = NULL,
  sexColours = c("red", "blue"),
  ...
)

fixPedigree(ped, dat = NULL)

genomesim(
  pedigree,
  founders = NULL,
  positions = NULL,
  initHe = NULL,
  mutationType = NULL,
  mutationRate = NULL,
  phenotyped = NULL,
  founderHaplotypes = NULL,
  genotyped = NULL,
  returnG = "n",
  initFreqs = NULL
)

microsim(
  pedigree,
  genFreqs = NULL,
  genotypesSample = NULL,
  knownGenotypes = NULL,
  records = NULL,
  eRate1 = 0,
  eRate2 = 0,
  eRate3 = 0
)

pedigreeStat()

pedStatSummary()

phensim(
  pedigree,
  traits = 1,
  randomA = NULL,
  randomE = NULL,
  parentalA = NULL,
  parentalE = NULL,
  sampled = NULL,
  records = NULL,
  returnAllEffects = FALSE
)
}
\value{
dependant on the deprecated function see pedantics help for details
}
\description{
The functions listed below are deprecated and are available for
backward compatibility only. They might be defunct in a future version
When possible, alternative functions with similar
functionality are also mentioned.
}
\section{\code{makePedigreeFactor}}{
 the function is now a special case of the \code{convert_ped} function when type is factor
}

\section{\code{makePedigreeNumeric}}{
 the function is now a special case of the \code{convert_ped} function when type is numeric
}

\section{\code{drawPedigree}}{
 the function has just been renamed with no other changes for the moment, but will soon be replace by "ggpedigree" function
}

\section{\code{fixPedigree}}{
 the function has just been renamed with no
other changes for the moment
}

\section{\code{genomesim}}{
 the function has just been renamed with no other changes for the moment
}

\section{\code{microsim}}{
 the function has just been renamed with no other changes for the moment
}

\section{\code{pedigreeStats}}{
 the function has been simplified but only the functionality are still available via \code{ped_stats} and its summary and plot methods
}

\section{\code{pedStatSummary}}{
 the function has been simplified but only the functionality are still available via \code{ped_stats} and its summary and plot methods
}

\section{\code{phensim}}{
 the function has just been renamed with no other changes for the moment
}

\keyword{internal}
