% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ped_stats_summary.R
\name{plot.ped_stats}
\alias{plot.ped_stats}
\title{Plot output from ped_stats}
\usage{
\method{plot}{ped_stats}(x, lowMem = FALSE, grContrast = FALSE, ...)
}
\arguments{
\item{x}{An object of class ped_stats generated by \code{ped_stats}}

\item{lowMem}{If TRUE, then stats based on calculation of A are not performed.}

\item{grContrast}{If TRUE, then uglier shades of red and blue are used to denote male and female statistics in graphical reports, but these colours provide better contrast in greyscale.}

\item{...}{extra arguments}
}
\value{
Returns a table of numbers of records, maternities,
paternities, pairwise sibship relationships, numbers
of different classes of grand-parental relationships,
pedigree depth, number of founders, mean subship sizes,
simple statistics of numbers of inbred and non-inbred
indivdiuals, and proportions of pairwise relationship
coefficients equal to or greater than several
thresholds.
}
\description{
Generates a manageable summary of pedigree-wide statistics
reported by ped_stats, either for a single pedigree or
for a comparison between pedigrees
}
\keyword{plot}
