% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ped_stats_summary.R
\name{summary.ped_stats}
\alias{summary.ped_stats}
\alias{pedStatSummary}
\title{Post-processes output from ped_stats}
\usage{
\method{summary}{ped_stats}(object, ...)

pedStatSummary(object, ...)
}
\arguments{
\item{object}{An object of class ped_stats generated by \code{ped_stats}}

\item{...}{extra arguments}
}
\value{
Returns a table of numbers of records, maternities,
paternities, pairwise sibship relationships, numbers
of different classes of grand-parental relationships,
pedigree depth, number of founders, mean subship sizes,
simple statistics of numbers of inbred and non-inbred
indivdiuals, and proportions of pairwise relationship
coefficients equal to or greater than several
thresholds.
}
\description{
Generates a manageable summary of pedigree-wide statistics
reported by ped_stats, either for a single pedigree or
for a comparison between pedigrees
}
