% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WarblerG}
\alias{WarblerG}
\title{Seychelles Warbler Genotypes}
\format{
An object of class \code{data.frame} with 307 rows and 29 columns.
}
\usage{
WarblerG
}
\description{
Genotype data collected by David Richardson from Cousin Island in 1999.
}
\references{
Richardson \emph{et.al.} (2001) Molecular Ecology 10 2263-2273
Hadfield J.D. \emph{et al} (2006) Molecular Ecology 15 3715-31
}
\keyword{dataset}
