% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpedigree.R
\name{ggpedigree}
\alias{ggpedigree}
\title{ggpedigree: Pedigree plotting tool for complex pedigrees.}
\usage{
ggpedigree(
  ids = NULL,
  mothers = NULL,
  fathers = NULL,
  cohort = NULL,
  sex = NULL,
  sex_female = NULL,
  sex_male = NULL,
  id_labels = FALSE,
  remove_singletons = TRUE,
  plot_unknown_cohort = TRUE,
  randomise_x_coordinates = FALSE,
  print_cohort_labels = TRUE,
  return_plot_tables = FALSE,
  suppress_plot = FALSE,
  line_col_mother = "#E41A1C",
  line_col_father = "#377EB8",
  line_alpha = 0.3,
  point_size = 1,
  point_colour = "black",
  point_alpha = 1,
  xlab = "",
  ylab = "",
  gg_theme = NULL
)
}
\arguments{
\item{ids}{A vector of individual identifiers}

\item{mothers}{A vector of mothers corresponding to ids. Missing values are 0
or NA.}

\item{fathers}{A vector of fathers corresponding to ids. Missing values are 0
or NA.}

\item{cohort}{integer. Default NULL. An optional vector assigning a cohort to
each id. If NULL, then \code{kinship2::kindepth} is used to assign cohorts to
ids.}

\item{sex}{integer or character. Default NULL. An optional vector assigning a
sex to each id. This can be any value, but the first level numerically or
alphabetically (e.g. 0 or "F") will be plotted with a circle (traditionally
denoting a female) and the second level will be plotted with a square
(traditionally denoting a male). Any negative values, NA values, or third
alphabetically values will be plotted with a triangle. NOTE: These can be
overridden by specifying \code{sex_female} or \code{sex_male} values.}

\item{sex_female}{integer or character. Default NULL. Indicates the value
used in \code{sex} for females. Will be plotted as circles.}

\item{sex_male}{integer or character. Default NULL. Indicates the value used
in \code{sex} for males. Will be plotted as squares.}

\item{id_labels}{logical. Default FALSE. Print the ids on the pedigree plot.}

\item{remove_singletons}{logical. Default TRUE. Remove ids with no relatives
i.e., no offspring or parents assigned.}

\item{plot_unknown_cohort}{logical. Default TRUE. Plots ids of unknown
cohorts. These are plotted in an "Unknown" cohort at the top of the
pedigree. Be aware that any mothers and fathers of these individuals will
be plotted below them.}

\item{randomise_x_coordinates}{logical. Default TRUE. Randomise the position of
individuals within each cohort.}

\item{print_cohort_labels}{logical. Default TRUE. Prints cohort levels on the
left hand side of plot.}

\item{return_plot_tables}{logical. Default FALSE. Returns an object with the
line and point data used for the plot.}

\item{suppress_plot}{logical. Default FALSE. Will stop the plot from being
output e.g., if using return_plot_tables to retrieve the plot objects.}

\item{line_col_mother}{Default = "#E41A1C". Line colour for maternal links.}

\item{line_col_father}{Default = "#377EB8". Line colour for paternal links.}

\item{line_alpha}{Default = 0.3. Line alpha (transparency) value for maternal
and paternal links.}

\item{point_size}{Default = 1. Point size for ids.}

\item{point_colour}{Default = "black". Point colour for ids.}

\item{point_alpha}{Default = 1. Point alpha for ids.}

\item{xlab}{Default = blank. x-axis label.}

\item{ylab}{Default = blank. y-axis label}

\item{gg_theme}{Alternate theme information specified as a ggplot::theme()
object.}
}
\description{
ggpedigree: Pedigree plotting tool for complex pedigrees.
}
\keyword{plot}
