% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedtricks-package.R
\docType{package}
\name{pedtricks-package}
\alias{pedtricks}
\alias{pedtricks-package}
\title{pedtricks: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Julien Martin \email{julien.martin@uottawa.ca} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

Authors:
\itemize{
  \item Matthew Wolak \email{mew0099@auburn.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})
  \item Susan Gordon \email{Susan.Johnston@ed.ac.uk} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})
}

}
